requirements.txt
selenium==4.19.0
webdriver-manager==4.0.1
requests==2.32.3
python-dotenv==1.0.1